+++ src/Http/Controllers/API/MatrixController.php
@@ -18,6 +18,8 @@ class MatrixController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('matrices.viewAny');
+
         $matrices = Matrix::orderBy('name')->paginate(25);
 
         return MatrixResource::collection($matrices);
@@ -31,6 +33,8 @@ class MatrixController extends Controller
      */
     public function show(Matrix $matrix)
     {
+        $this->authorize('matrices.view');
+
         return new MatrixResource($matrix);
     }
 
@@ -42,6 +46,8 @@ class MatrixController extends Controller
      */
     public function slug($matrix)
     {
+        $this->authorize('matrices.view');
+
         $matrix = Matrix::where('slug', $matrix)->firstOrFail();
 
         return new MatrixResource($matrix);
@@ -92,6 +98,8 @@ class MatrixController extends Controller
      */
     public function destroy(Matrix $matrix)
     {
+        $this->authorize('matrices.delete');
+
         $matrix->delete();
     }
 }
